Remove diacritics prior to running StringScore matching
This allows the search function to see "Jalapeño" and "Jalapeno" as exactly the same word, important given that users on mobile are unlikely to ever take the time to include the diacritics, and that there are a number of events and volunteers in this year's programme that ask for this.

Removal of the diacritics is achieved by normalizing to the string's canonical decomposition, and then removing the second bytes (in range of \u0300 - \u036f) from the composition. This functionality isn't available in Internet Explorer 11, so fall back to normalizing casing only when the functionality is not available.

Largely by virtue of executing a regular regression, this does increase the StringScore cost from 2.09 to 8.05 microseconds (+3.85x). Given an event programme of ~500 entities, that brings comparison time to just over 4 milliseconds, still well within our frame budget.